Refactor trait impls in `spinoso-array`
Remove some trait impls and move `IntoIterator` impls to an `iter`
module in each array subtype.

This commit removes the following impls:

- `AsRef<Vec<T>> for Array`
- `AsMut<Vec<T>> for Array`
- `AsRef<SmallVec<[T; N]>> for SmallArray`
- `AsMut<SmallVec<[T; N]>> for SmallArray`
- `AsRef<TinyVec<[T; N]>> for TinyArray`
- `AsMut<TinyVec<[T; N]>> for TinyArray`

This commit addresses a namespace conflict with `core::iter` by
reimplementing `TinyArray::with_len_and_default` using `TinyVec::new`
and `TinyVec::resize`.

Removing trait impls is a breaking change, even though they are unused
in `artichoke-backend`. Bump `spinoso-array` to 0.8.0 and update README.
