+++ artichoke-backend/Cargo.toml
@@ -25,7 +25,7 @@ onig = { version = "6.3.0", optional = true, default-features = false }
 # https://github.com/artichoke/artichoke/pull/1729
 regex = "1.5.5"
 scolapasta-string-escape = { version = "0.2.0", path = "../scolapasta-string-escape", default-features = false }
-spinoso-array = { version = "0.7.0", path = "../spinoso-array", default-features = false }
+spinoso-array = { version = "0.8.0", path = "../spinoso-array", default-features = false }
 spinoso-env = { version = "0.1.1", path = "../spinoso-env", optional = true, default-features = false }
 spinoso-exception = { version = "0.1.0", path = "../spinoso-exception" }
 spinoso-math = { version = "0.2.0", path = "../spinoso-math", optional = true, default-features = false }
