+++ spinoso-array/src/array/smallvec/impls.rs
@@ -1,30 +1,13 @@
 use core::borrow::{Borrow, BorrowMut};
 use core::ops::{Deref, DerefMut, Index, IndexMut};
-use core::slice::{Iter, IterMut, SliceIndex};
+use core::slice::SliceIndex;
 
-use smallvec::SmallVec;
-
-use crate::array::smallvec::SmallArray;
-use crate::array::INLINE_CAPACITY;
-
-impl<T> AsRef<SmallVec<[T; INLINE_CAPACITY]>> for SmallArray<T> {
-    #[inline]
-    fn as_ref(&self) -> &SmallVec<[T; INLINE_CAPACITY]> {
-        &self.0
-    }
-}
+use super::SmallArray;
 
 impl<T> AsRef<[T]> for SmallArray<T> {
     #[inline]
     fn as_ref(&self) -> &[T] {
         self.0.as_ref()
-    }
-}
-
-impl<T> AsMut<SmallVec<[T; INLINE_CAPACITY]>> for SmallArray<T> {
-    #[inline]
-    fn as_mut(&mut self) -> &mut SmallVec<[T; INLINE_CAPACITY]> {
-        &mut self.0
     }
 }
 
@@ -101,25 +84,5 @@ where
     #[inline]
     fn index_mut(&mut self, index: I) -> &mut I::Output {
         &mut self.0[index]
-    }
-}
-
-impl<'a, T> IntoIterator for &'a SmallArray<T> {
-    type Item = &'a T;
-    type IntoIter = Iter<'a, T>;
-
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.iter()
-    }
-}
-
-impl<'a, T> IntoIterator for &'a mut SmallArray<T> {
-    type Item = &'a mut T;
-    type IntoIter = IterMut<'a, T>;
-
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.iter_mut()
     }
 }
