+++ spinoso-array/src/array/smallvec/iter.rs
@@ -1,0 +1,23 @@
+use core::slice::{Iter, IterMut};
+
+use super::SmallArray;
+
+impl<'a, T> IntoIterator for &'a SmallArray<T> {
+    type Item = &'a T;
+    type IntoIter = Iter<'a, T>;
+
+    #[inline]
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.iter()
+    }
+}
+
+impl<'a, T> IntoIterator for &'a mut SmallArray<T> {
+    type Item = &'a mut T;
+    type IntoIter = IterMut<'a, T>;
+
+    #[inline]
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.iter_mut()
+    }
+}
